Refactor shared logic out
