+++ buildSrc/src/main/java/com/google/firebase/gradle/plugins/FirebaseJavaLibraryPlugin.kt
@@ -4,13 +4,13 @@ import com.github.sherter.googlejavaformatgradleplugin.GoogleJavaFormatExtension
 import com.google.common.collect.ImmutableList
 import com.google.common.collect.ImmutableMap
 import java.io.File
-import java.nio.file.Paths
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.api.attributes.Attribute
 import org.gradle.api.plugins.JavaPluginConvention
 import org.gradle.api.publish.PublishingExtension
 import org.gradle.api.publish.maven.MavenPublication
+import org.gradle.kotlin.dsl.getPlugin
 import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
 
 class FirebaseJavaLibraryPlugin : Plugin<Project> {
@@ -38,40 +38,13 @@ class FirebaseJavaLibraryPlugin : Plugin<Project> {
   }
 
   private fun setupApiInformationAnalysis(project: Project) {
-    val mainSourceSet =
-      project.convention.getPlugin(JavaPluginConvention::class.java).sourceSets.getByName("main")
-    val outputFile =
-      project.rootProject.file(
-        Paths.get(
-          project.rootProject.buildDir.path,
-          "apiinfo",
-          project.path.substring(1).replace(":", "_")
-        )
-      )
-    val outputApiFile = File(outputFile.absolutePath + "_api.txt")
-    val apiTxt =
-      if (project.file("api.txt").exists()) project.file("api.txt")
-      else project.file(project.rootDir.toString() + "/empty-api.txt")
-    val apiInfo =
-      project.tasks.register("apiInformation", ApiInformationTask::class.java) {
-        sources.value(project.provider { mainSourceSet.java.srcDirs })
-        apiTxtFile.set(apiTxt)
-        baselineFile.set(project.file("baseline.txt"))
-        this.outputFile.set(outputFile)
-        this.outputApiFile.set(outputApiFile)
-        updateBaseline.set(project.hasProperty("updateBaseline"))
-      }
-    val generateApiTxt =
-      project.tasks.register("generateApiTxtFile", GenerateApiTxtTask::class.java) {
-        sources.value(project.provider { mainSourceSet.java.srcDirs })
-        apiTxtFile.set(project.file("api.txt"))
-        baselineFile.set(project.file("baseline.txt"))
-        updateBaseline.set(project.hasProperty("updateBaseline"))
-      }
-    val docStubs =
-      project.tasks.register("docStubs", GenerateStubsTask::class.java) {
-        sources.value(project.provider { mainSourceSet.java.srcDirs })
-      }
+    val srcDirs =
+      project.convention.getPlugin<JavaPluginConvention>().sourceSets.getByName("main").java.srcDirs
+
+    val apiInfo = getApiInfo(project, srcDirs)
+    val generateApiTxt = getGenerateApiTxt(project, srcDirs)
+    val docStubs = getDocStubs(project, srcDirs)
+
     project.tasks.getByName("check").dependsOn(docStubs)
     project.afterEvaluate {
       val classpath =
@@ -88,6 +61,7 @@ class FirebaseJavaLibraryPlugin : Plugin<Project> {
       docStubs.configure { classPath = classpath }
     }
   }
+
   private fun configurePublishing(project: Project, firebaseLibrary: FirebaseLibraryExtension) {
     project.apply(ImmutableMap.of("plugin", "maven-publish"))
     val publishing = project.extensions.getByType(PublishingExtension::class.java)
